[#2] refactor: Post 모델 이동
