--- src/models/post.ts
@@ -1,6 +1,0 @@
-import { MdContent } from "@/models/markdown-content";
-
-/**
- * 포스트
- */
-export interface Post extends MdContent {}
